==> 1_teams_data.py <==
# -*- coding: utf-8 -*-
"""
Created on Wed Apr  2 11:33:17 2025

@author: julia
"""

import pandas as pd
import geopandas as gpd

# Read the file with team data, including locations

teams = gpd.read_file("./inputs/Major_League_Baseball_Stadiums-shp.zip")

# Drop unnecessary columns

drop = ["OBJECTID","CITY","STATE","CAPACITY"]
teams = teams.drop(columns=drop)

# Read in the information about abbreviations and divisions
# The divs_abb.csv file was generated by me based on information from Baseball Reference

divs_abb = pd.read_csv("./inputs/divs_abb.csv")

# Merge the abbreviations data onto the team data. 

teams = teams.merge(divs_abb,how="left",on="TEAM")

# Define the ouptputs folder. Save to a csv in the folder

output_folder = './outputs/'
teams.to_csv(output_folder + "teams.csv")

# Create a dataframe of locations only

loc_drop = ['NAME', 'LEAGUE', 'TEAM']
locations = teams.drop(columns=loc_drop)
locations["location"] = locations["ABBREVIATION"]  

# Save to a pickle file in the outputs folder

locations.to_pickle(output_folder + "team_locations.pkl")
